move createFakeModuleFSWithManifest to its own test helper file
